EES-3998 - switch to using new create snapshots scripts
